requirements.txt
asgiref==3.7.1
backend==0.2.4.1
backends==1.5.3
backports.zoneinfo==0.2.1
beartype==0.14.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2.1
django-oauth-plus==2.2.9
django-oauth-toolkit==2.2.0
django-oauth2-provider==0.2.6.1
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
fdm==0.4.1
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
jwcrypto==1.4.2
mpmath==1.3.0
numpy==1.24.3
oauth2==1.9.0.post1
oauthlib==3.2.2
opt-einsum==3.3.0
plum-dispatch==2.0.1
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.0.9
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
sympy==1.12
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==1.26.16
wrapt==1.15.0
